Remove sexism from README.md
I do not believe that any harm was intended behind this comment, but
a seemingly harmless joke such as women preferring pinks and purples
because society has deemed them to be "girly" colors is more likely to
be a bad thing. It is an example of "special" treatment in a field where
women do not want to be treated differently for being women.
Additionally, it perpetuates the idea that pinks and purples can not be
for men. But come on, guys: who doesn't like purple? Seriously.

Signed-off-by: David Celis <me@davidcel.is>